+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Move tests from Pester to mocha
 * Move from TravisCI to Github Actions
 * Tweak publishing workflow
+* Improve version parsing script to pull and parse all releases from Github's API.
+* Add many more pwsh versions, including arm and arm64
 
 # v0.2.0